[packages] php5: use bundled libgd for gd extension
PHP devs doesn't only include an own copy of libgd but modified
it. As long as these modifications are not part of the offical
libgd distribution, we have to deal with the bundled version.

Details can be found at 
https://bugs.php.net/bug.php?id=55224 and
https://bugs.php.net/bug.php?id=60108

When this problem is fixed upstream, this patch should be reverted.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>



git-svn-id: svn://svn.openwrt.org/openwrt/packages@32879 3c298f89-4303-0410-b956-a3cf2f4a3e73