isc-dhcp-relay: Add init scripts for ipv4 and ipv6 as well as config file
I cleaned up the patch so line do not overflow and removed some white
space in hope to get it accepted or revised.

As isc-dhcp-relay does not come with an init script, I made one based
on sysntpd init.d file for both ipv4 and ipv6.

If you need more arguments from isc-dhcp-relay to be included in the
config file, I'll be able to add them.

This patch has been applied to trunk and compiles/install on my side.

The script for ipv4 and ipv6 have been tested with the
isc-dhcp-relay-ipv6 binary and work well (start/stop with the right
settings).

Signed-off-by: Jérôme Poulin <jeromepoulin@gmail.com>
Signed-off-by: Florian Fainelli <florian@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@34485 3c298f89-4303-0410-b956-a3cf2f4a3e73