rygel: add package
Plugins other than playbin are untested.

Rygel is normally run as a desktop service under a user account. It
doesn't behave like a system daemon. Since running as a desktop service
doesn't make sense for OpenWrt, I've included an init script and made a
crude change to send the output to syslog.

Signed-off-by: Andy Leiserson <andy@leiserson.org>
Signed-off-by: Florian Fainelli <florian@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@34528 3c298f89-4303-0410-b956-a3cf2f4a3e73