ffmpeg: version bump to 0.11.2
Update FFmpeg to version 0.11.2. The API changed from 0.8. Minidlna in the
feeds works with this, and a patch is forthcoming for libdlna. I'm not aware
of what other packages might break with this update.

Two new ffmpeg tools, aviocat and ffeval, are not included in the package.
ffeval is used for their development tests. Aviocat is used for dumping
media streams.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@35795 3c298f89-4303-0410-b956-a3cf2f4a3e73