[packages] wview: fix operation for low temperatures for WMRUSB
When the outside temperature is less than -10C the station reports that
dewpoint is invalid. This patch allows to ignore only that value instead
of ignoring all of the packet contents from the outside sensor (including
temperature and humidity).

This is also sort of fixed (the reported dew point values might be
invalid) in upstream release 5.20.0, but we can't use his repository
because there is no public repository and also he ignores requests on
his semi-private bugtracker; so we need to use the source we can
trust, and currently jgoerzen's branch is the best source out there.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>

git-svn-id: svn+ssh://svn.openwrt.org/openwrt/packages@36389 3c298f89-4303-0410-b956-a3cf2f4a3e73