packages: add mosh
This adds mosh-full, mosh-server, and mosh-client packages.

The Entware team has already done the hard work:
http://code.google.com/p/wl500g-repo/source/browse/feeds/rtndev/mosh/Makefile?r=1220

All I added was a patch to disabled the check for UTF-8, which is not
available on our platform. (It works fine with the check disabled)
https://github.com/keithw/mosh/issues/464

This was originally requested here:
https://dev.openwrt.org/ticket/12509

This actually does work in conjunction with the openssh-client/server.
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=662939

Signed-off-by: Kyle Anderson <kyle@xkyle.com>


git-svn-id: svn://svn.openwrt.org/openwrt/packages@39541 3c298f89-4303-0410-b956-a3cf2f4a3e73