luaposix: Update source links and urls
No functional change.  The difference in the md5sum of the tarball is
the inclusion of a .gitignore file in the github provided tarball.

This change allows https://home.comcast.net/~sdwalker/uscan/index.html
to more accurately track the state of this package, and includes links
that actually work.  It eliminates the need for this package to be
hosted on mirror2.openwrt.org, and indeed, due to the md5sum change,
won't even be compatible with the package hosted there.

Signed-off-by: Karl Palsson <karlp@remake.is>


git-svn-id: svn://svn.openwrt.org/openwrt/packages@40611 3c298f89-4303-0410-b956-a3cf2f4a3e73